destination-e2e: merge cloud and OSS
