+++ airbyte-integrations/connectors/destination-dev-null/src/main/kotlin/io/airbyte/integrations/destination/dev_null/SilentDestination.kt
@@ -1,23 +1,19 @@
 /*
  * Copyright (c) 2023 Airbyte, Inc., all rights reserved.
  */
-package io.airbyte.integrations.destination.e2e_test
+package io.airbyte.integrations.destination.dev_null
 
 import com.fasterxml.jackson.databind.JsonNode
 import io.airbyte.cdk.integrations.BaseConnector
 import io.airbyte.cdk.integrations.base.AirbyteMessageConsumer
 import io.airbyte.cdk.integrations.base.Destination
-import io.airbyte.integrations.destination.e2e_test.logging.LoggingConsumer
-import io.airbyte.integrations.destination.e2e_test.logging.TestingLoggerFactory
 import io.airbyte.protocol.models.v0.AirbyteConnectionStatus
 import io.airbyte.protocol.models.v0.AirbyteMessage
 import io.airbyte.protocol.models.v0.ConfiguredAirbyteCatalog
 import java.util.function.Consumer
-import org.slf4j.Logger
-import org.slf4j.LoggerFactory
 
-/** This destination logs each record it receives. */
-class LoggingDestination : BaseConnector(), Destination {
+/** This destination silently receives records. */
+class SilentDestination : BaseConnector(), Destination {
     override fun check(config: JsonNode): AirbyteConnectionStatus {
         return AirbyteConnectionStatus().withStatus(AirbyteConnectionStatus.Status.SUCCEEDED)
     }
@@ -27,10 +23,19 @@ class LoggingDestination : BaseConnector(), Destination {
         catalog: ConfiguredAirbyteCatalog,
         outputRecordCollector: Consumer<AirbyteMessage>
     ): AirbyteMessageConsumer {
-        return LoggingConsumer(TestingLoggerFactory(config), catalog, outputRecordCollector)
+        return RecordConsumer(outputRecordCollector)
     }
 
-    companion object {
-        private val LOGGER: Logger = LoggerFactory.getLogger(LoggingDestination::class.java)
+    class RecordConsumer(private val outputRecordCollector: Consumer<AirbyteMessage>) :
+        AirbyteMessageConsumer {
+        override fun start() {}
+
+        override fun accept(message: AirbyteMessage) {
+            if (message.type == AirbyteMessage.Type.STATE) {
+                outputRecordCollector.accept(message)
+            }
+        }
+
+        override fun close() {}
     }
 }
