+++ airbyte-integrations/connectors/destination-dev-null/src/main/kotlin/io/airbyte/integrations/destination/dev_null/TestingDestinations.kt
@@ -1,17 +1,22 @@
 /*
  * Copyright (c) 2023 Airbyte, Inc., all rights reserved.
  */
-package io.airbyte.integrations.destination.e2e_test
+package io.airbyte.integrations.destination.dev_null
 
 import com.fasterxml.jackson.databind.JsonNode
+import com.fasterxml.jackson.databind.node.ArrayNode
+import com.fasterxml.jackson.databind.node.ObjectNode
 import com.google.common.collect.ImmutableMap
 import io.airbyte.cdk.integrations.BaseConnector
 import io.airbyte.cdk.integrations.base.AirbyteMessageConsumer
 import io.airbyte.cdk.integrations.base.Destination
 import io.airbyte.cdk.integrations.base.IntegrationRunner
+import io.airbyte.commons.features.EnvVariableFeatureFlags
+import io.airbyte.commons.features.FeatureFlags
 import io.airbyte.protocol.models.v0.AirbyteConnectionStatus
 import io.airbyte.protocol.models.v0.AirbyteMessage
 import io.airbyte.protocol.models.v0.ConfiguredAirbyteCatalog
+import io.airbyte.protocol.models.v0.ConnectorSpecification
 import java.util.function.Consumer
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
@@ -19,13 +24,14 @@ import org.slf4j.LoggerFactory
 class TestingDestinations
 @JvmOverloads
 constructor(
+    override val featureFlags: FeatureFlags = EnvVariableFeatureFlags(),
     private val destinationMap: Map<TestDestinationType, Destination> =
         ImmutableMap.builder<TestDestinationType, Destination>()
             .put(TestDestinationType.LOGGING, LoggingDestination())
             .put(TestDestinationType.THROTTLED, ThrottledDestination())
             .put(TestDestinationType.SILENT, SilentDestination())
             .put(TestDestinationType.FAILING, FailAfterNDestination())
-            .build()
+            .build(),
 ) : BaseConnector(), Destination {
     enum class TestDestinationType {
         LOGGING,
@@ -40,6 +46,33 @@ constructor(
                 config["test_destination"]["test_destination_type"].asText()
             )
         ]
+    }
+
+    override fun spec(): ConnectorSpecification {
+        if (!isCloudDeployment) {
+            return super.spec()
+        } else {
+            /** 1. Update the title. 2. Only keep the "silent" mode. */
+            val spec = super.spec()
+
+            (spec.connectionSpecification as ObjectNode).put("title", DEV_NULL_DESTINATION_TITLE)
+
+            val properties =
+                spec.connectionSpecification["properties"]["test_destination"] as ObjectNode
+            val types = properties["oneOf"] as ArrayNode
+            val typesIterator = types.elements()
+            while (typesIterator.hasNext()) {
+                val typeNode = typesIterator.next()
+                if (
+                    !typeNode["properties"]["test_destination_type"]["const"]
+                        .asText()
+                        .equals("silent", ignoreCase = true)
+                ) {
+                    typesIterator.remove()
+                }
+            }
+            return spec
+        }
     }
 
     @Throws(Exception::class)
@@ -64,6 +97,7 @@ constructor(
 
     companion object {
         private val LOGGER: Logger = LoggerFactory.getLogger(TestingDestinations::class.java)
+        private const val DEV_NULL_DESTINATION_TITLE = "E2E Test (/dev/null) Destination Spec"
 
         @Throws(Exception::class)
         @JvmStatic
