+++ airbyte-integrations/connectors/destination-dev-null/src/main/kotlin/io/airbyte/integrations/destination/dev_null/logging/EveryNthLogger.kt
@@ -1,34 +1,33 @@
 /*
  * Copyright (c) 2023 Airbyte, Inc., all rights reserved.
  */
-package io.airbyte.integrations.destination.e2e_test.logging
+package io.airbyte.integrations.destination.dev_null.logging
 
 import io.airbyte.protocol.models.v0.AirbyteRecordMessage
 import io.airbyte.protocol.models.v0.AirbyteStreamNameNamespacePair
-import java.util.*
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 
-class RandomSamplingLogger(
+class EveryNthLogger(
     streamNamePair: AirbyteStreamNameNamespacePair,
-    private val samplingRatio: Double,
-    seed: Long,
+    private val nthEntryToLog: Int,
     maxEntryCount: Int
 ) : BaseLogger(streamNamePair, maxEntryCount), TestingLogger {
-    private val random = Random(seed)
+    private var currentEntry = 0
 
     override fun log(recordMessage: AirbyteRecordMessage?) {
         if (loggedEntryCount >= maxEntryCount) {
             return
         }
 
-        if (random.nextDouble() < samplingRatio) {
+        currentEntry += 1
+        if (currentEntry % nthEntryToLog == 0) {
             loggedEntryCount += 1
             LOGGER.info(entryMessage(recordMessage!!))
         }
     }
 
     companion object {
-        private val LOGGER: Logger = LoggerFactory.getLogger(RandomSamplingLogger::class.java)
+        private val LOGGER: Logger = LoggerFactory.getLogger(EveryNthLogger::class.java)
     }
 }
