+++ airbyte-integrations/connectors/destination-dev-null/src/main/resources/spec.json
