+++ airbyte-integrations/connectors/destination-dev-null/src/test-integration/kotlin/io/airbyte/integrations/destination/dev_null/CloudDevNullDestinationAcceptanceTest.kt
@@ -1,26 +1,25 @@
 /*
  * Copyright (c) 2023 Airbyte, Inc., all rights reserved.
  */
-package io.airbyte.integrations.destination.e2e_test
+package io.airbyte.integrations.destination.dev_null
 
 import com.fasterxml.jackson.databind.JsonNode
-import io.airbyte.cdk.integrations.standardtest.destination.DestinationAcceptanceTest
 import io.airbyte.commons.json.Jsons
-import io.airbyte.integrations.destination.e2e_test.TestingDestinations.TestDestinationType
-import io.airbyte.protocol.models.v0.*
+import io.airbyte.protocol.models.v0.AirbyteMessage
+import io.airbyte.protocol.models.v0.AirbyteRecordMessage
 import java.util.*
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.Test
 
-class TestingSilentDestinationAcceptanceTest : DestinationAcceptanceTest() {
-    override val imageName = "airbyte/destination-e2e-test:dev"
+class CloudDevNullDestinationAcceptanceTest : TestingSilentDestinationAcceptanceTest() {
+    override val isCloudTest = true
 
     override fun getConfig(): JsonNode {
         return Jsons.jsonNode(
             Collections.singletonMap(
                 "test_destination",
-                Collections.singletonMap("test_destination_type", TestDestinationType.SILENT.name)
+                Collections.singletonMap("test_destination_type", "SILENT")
             )
         )
     }
@@ -59,21 +58,12 @@ class TestingSilentDestinationAcceptanceTest : DestinationAcceptanceTest() {
         Assertions.assertEquals(0, actual.size)
     }
 
-    open override fun getDefaultSchema(config: JsonNode): String? {
-        return super.getDefaultSchema(config) ?: "default_schema"
-    }
-
     // Skip because `retrieveRecords` returns an empty list at all times.
     @Disabled @Test override fun testSyncNotFailsWithNewFields() {}
 
     @Disabled @Test override fun testAirbyteTimeTypes() {}
 
-    // This test assumes that dedup support means normalization support.
-    // Override it to do nothing.
-    @Disabled
-    @Test
-    @Throws(Exception::class)
-    override fun testIncrementalDedupeSync() {
-        super.testIncrementalDedupeSync()
+    open override fun getDefaultSchema(config: JsonNode): String? {
+        return super.getDefaultSchema(config) ?: "default_schema"
     }
 }
