+++ airbyte-integrations/connectors/destination-dev-null/src/test/resources/expected_spec_cloud.json
