--- airbyte-integrations/connectors/destination-e2e-test/build.gradle
@@ -1,17 +1,0 @@
-plugins {
-    id 'airbyte-java-connector'
-}
-
-airbyteJavaConnector {
-    cdkVersionRequired = '0.45.0'
-    features = ['db-destinations']
-    useLocalCdk = true
-}
-
-application {
-    mainClass = 'io.airbyte.integrations.destination.e2e_test.TestingDestinations'
-    applicationDefaultJvmArgs = ['-XX:+ExitOnOutOfMemoryError', '-XX:MaxRAMPercentage=75.0']
-}
-
-dependencies {
-}
